Som användare vill jag att rätt QR-kod från mitt senaste köp visas (#86)
* Every receipt now has a property 'active'
This property is used to tell if a receipt is active or not, and so decides if the QR code has been scanned already

* Började med att få rätt kvitto att synas och tog bort CafeCheckoutButton, som inte användes längre

* Setup endpoint for invalidating a receipt
this endpoint expects a receipt id as a paremeter. It will then request firebase to update the given receipt's active status

* Refactoring: DRY
applied boy scout rule and refactored common behavior among functions communicating with firebase

* Refactor: How to fetch a set of receipts

* Invalidate now checks if receipt already exists
This is because if we try to update an object that does not exists with firebase's REST-API, then firebase will create that object for us. Not good ..

* Refactor: clarified log message

* Kopplat så att beställningen man lägger nu läggs in i QR-koden

* Shop is now being resolved in the background when adding coffees to the cart

* Added reducer for adding a shop to the cart

* Refactored resolving shop function

* Receipts active property is now set to false the first time they are fetched

* no qr active screen

* Added new endpoint: scanReceipt
scanReceipt will both invalidate and return the receipt passed to it, 'simulating' a scan of the QR-code

* Now logging wether the scanned receipt is valid or not

* sendOrder now returns a receipt id instead of a link to a receipt

* re-added link to receipt when scanning qr-code

* Added function in API for retrieving a link for 'scanning' a receipt